% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/functions.R
\name{buildKeywordList}
\alias{buildKeywordList}
\title{This function generates keyword and ignore lists based on the expansion of
concepts.}
\usage{
buildKeywordList(connection, aphroditeConceptName, schema, dbms)
}
\arguments{
\item{connection}{The connection to the database server.}

\item{aphroditeConceptName}{The string of text / concept name to use.}

\item{schema}{The database schema being used.}

\item{dbms}{The target DBMS for SQL to be rendered in.}
}
\value{
A list with two elements: a list of positive keywords found
(keywordlist_ALL), and a list of ignore keywords (ignorelist_ALL)
}
\description{
Given any given concept_id or string of text this function
generates keyword and ignore lists based on the expansion of concepts (looking
at their synonyms).
}
\details{
Takes the aphroditeConceptName looks for synonyms and builds a list
of related concepts using the vocabulary hierarchies
}
\examples{
\dontrun{

wordLists <- buildKeywordList(conn, aphrodite_concept_name, cdmSchema, dbms)

}
}

