% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{buildModel}
\alias{buildModel}
\title{This function builds a model for the specified feature vector using cases and
controls for a certain outcomeName}
\usage{
buildModel(flags, pp_total, outcomeNameS, saveFolder)
}
\arguments{
\item{flags}{The R dataframe that contains all feature/model flags
specified in settings.R.}

\item{outcomeNameS}{String description of the outcome for which the model is
being built}

\item{saveFolder}{folder in which summary file output will be saved}

\item{featureVector}{Flattened feature vector returned by combineFeatureVectors
function, with labeled cases and controls.  Assumed to have one column named "Class_labels"
and one named "pid"}
}
\value{
An transferable caret Model object
}
\description{
This function builds a model for the specified feature vector
using cases and controls for a certain outcomeName. Returns a caret trained
model.
}
\details{
This function builds a model for the specified outcomeName. The
model is specified in the flags dataframe (currently only supports LASSO).
}
\examples{
\dontrun{

 model_predictors <- buildModel(flag, fv_all, predictorsNames, outcomeName, saveFolder)

}

}

