% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{combineFeatureVectors}
\alias{combineFeatureVectors}
\title{This function combines all of the desired feature types into one single feature vector}
\usage{
combineFeatureVectors(flags, cases_pids, controls_pids, featureVector,
  outcomeNameS)
}
\arguments{
\item{flags}{The R dataframe that contains all feature/model flags
specified in settings.R.}

\item{cases_pids}{List of patient_id's considered cases (for labeling
purposes)}

\item{controls_pids}{List of patient_id's considered controls (for labeling
purposes)}

\item{featureVector}{List of flattened feature vectors returned by buildFeatureVector
function.}

\item{outcomeNameS}{String description of the outcome for which the model is
being built [Not actually needed]}
}
\value{
fv_all - The combined feature vector (n patients x n features).  The columns are: pid column, predictorNames, outcomeName
}
\description{
This function combines all of the desired feature types into one single feature vector.  This feature vector is ready to be used for training
}
\details{
This function builds a feature vector by concatenating all of the available datasets.
If binary features are specified in the settings, this conversion is made.
The cases_pids and control_pids are patient_id's used for the labeling of
the testing and training sets.
}
\examples{
\dontrun{

 fv_full_data <- combineFeatureVectors(flag, cases, controls, fv_all, outcomeName)

}

}

