% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{conceptDecoder}
\alias{conceptDecoder}
\title{This function returns the concept terms corresponding to an input set of concept
IDs.}
\usage{
conceptDecoder(connection, schema, dbms, model, numFeats, breaker = ":",
  typeInd = 1, idInd = 2)
}
\arguments{
\item{connection}{The connection to the database server.}

\item{schema}{The database schema being used}

\item{dbms}{The target DBMS for SQL to be rendered in.}

\item{model}{The model object; will be used to extract top-ranking features}

\item{numFeats}{The number of features you'd like returned}

\item{breaker=":"}{Which sort of breaker is used in feature names (e.g. for "obs:12345" it would be ":")}

\item{typeInd1=1}{Indice after string split defining which feature class (e.g. [1] for "obs:12345" defines obs)}

\item{idInd=1}{Indice after string split defining which concept_id (e.g. [2] for "obs:12345
defines "12345")}
}
\value{
A list of concept terms and concept ids, corresponding to the IDs of interest
}
\description{
This function returns the concept terms corresponding to an input set of concept
IDs.
}
\details{
This function returns the concept terms corresponding to an input set of concept
IDs.  Use case: to investigate highly-ranked features from classification model
}
\examples{
\dontrun{

 high_ranking_concepts <- conceptDecoder(connection, schema, dbms, model, 20)

}

}

