% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{convertFeatVecPortion}
\alias{convertFeatVecPortion}
\title{This function builds a feature vector for a specific subset of features}
\usage{
convertFeatVecPortion(featuresType, key, labIndic)
}
\arguments{
\item{featuresType}{A set of patient data in the form of a list of data frames.  Each data frame contains a pid to label the patient, the names of the features that the patient had present, and the frequency counts of these features in his/her record}

\item{key}{String descriptor of type of feature (e.g. "obs:" or "visit:"). This will be used to label the feature}

\item{labIndic=0}{Whether this is for a lab feature.  If so, must be converted from factor to numeric.  Default is 0=no conversion required; 1=conversion required.}
}
\value{
An data frame of (pts) x (features of input type)
}
\description{
This function builds a feature matrix for a specific subset of features, e.g. labs/visits/observations/drug exposures.
Returns a feature matrix with all features from all patients included.
}
\details{
This function takes a list of patient data frames as input.  Each patient's data frame contains the features that this patient has present in his/her record. This function flattens this information into the combined feature matrix, with all features (of a certain type - e.g. labs or visits) from all patients included.  Clearly, many patients will not have data for many features; their feature counts for any feature that was not present in their record will be set as 0.
}
\examples{
\dontrun{

 FV_converted<-convertFeatVecPortion(featuresType, 'obs:')

 #OR

 FV_converted<-convertFeatVecPortion(featuresType, 'labs:', labIndic=1)

}

}
