% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/functions.R
\name{executeSQL}
\alias{executeSQL}
\title{This function executes one single SQL statement}
\usage{
executeSQL(connection, schema, query, targetDBMS)
}
\arguments{
\item{connection}{The connection to the database server.}

\item{schema}{The database schema being used.}

\item{query}{The SQL statement to retrieve the data.}

\item{targetDBMS}{The target DBMS for SQL to be rendered in.}
}
\value{
An object containing the data.
}
\description{
This function renders, translates and executes one single SQL
  statement that produces a result
}
\details{
Renders, translates, and executes a single SQL statement that is
  expeting to produce and result.
}
\examples{
\dontrun{

  library("SqlRender")
  library("DatabaseConnector")
  library("Aphordite")
  connectionDetails <- createConnectionDetails(dbms="mysql", server="localhost",
         user="root", password="blah" ,schema="cdm_v5")
  conn <- connect(connectionDetails)

  concept_of_interest <- executeSQL(connection, schema, paste("SELECT concept_id,
         concept_name FROM @cdmSchema.concept WHERE lower(concept_name) =
         lower('myocardial infarction') AND standard_concept = 'S' AND
         invalid_reason IS NULL AND domain_id = 'Condition';" ,sep = ""),dbms)

  dbDisconnect(conn)
}
}

