% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{f_score_calc}
\alias{f_score_calc}
\title{This function creates a summary metric for model training}
\usage{
f_score_calc(data, lev = levels(data$obs), model = NULL)
}
\arguments{
\item{data}{A dataframe of the held-out example cases, with columns for 'obs', 'pred', 'T', 'F'.  'T' and 'F' have the probabilities of each of these classes}

\item{lev}{Outcome factor levels for model}

\item{model}{Character string of model used}
}
\value{
f_score
}
\description{
This function creates a new summary metric for model training, specific for unbalanced classes.  Inputs as specified in caret.
}
\details{
This function returns the F-score for model training optimization.  Beta is currently set at 2 - TODO: should make this edit-able in future version.
}
\examples{
\dontrun{

 f_score <- f_score_calc(data, lev, model)

}
}

