% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{getAnchors}
\alias{getAnchors}
\title{This function allows Anchor recommendation}
\usage{
getAnchors(connection, dbms, schema, casesList, controlsList, ignores,
  studyName, outcomeName, flag, numAnchors)
}
\arguments{
\item{connection}{The connection to the database server.}

\item{dbms}{The target DBMS for SQL to be rendered in.}

\item{schema}{The database schema being used.}

\item{casesList}{The list of case patients (already filtered by keywords or gold standard).}

\item{controlsList}{The list of control patients.}

\item{ignores}{The list of concept_id's ignored when building the cohort.}

\item{studyName}{The study name(will be used for file naming).}

\item{outcomeName}{The outcomeName (will be use for modeling).}

\item{flag}{The flags variable containg the study configuration - we
use this one here to have flexibility of having two different sets of settings
for the same experimental run.}

\item{numAnchors}{The total number of anchors to be returned (top N features).}
}
\value{
A list of anchors containing rank, conceptID, domaiID
}
\description{
This function allows Anchor recommendation based after your initial set
of keyword and ignore lists have been provided. This will help improve model by
suggesting related features that were not considered initially.
}
\details{
This function takes the lists of exclude keywords and fetches all patient data
for the patients on the cases and controls list. It then builds a model to identify the
top performing features and returns a list of them as anchors. This new keyword list can
be feed to the set of Anchors specific functions to use any anchor as a selection criteria
for patients
}
\examples{
\dontrun{

numAnchors<-50
anchor_list <- getAnchors(conn, dbms, cdmSchema, cases, controls, as.character(ignoreList_FF$V3), studyName, outcomeName, flag, numAnchors)

}
}

