% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getNormalizationTerm}
\alias{getNormalizationTerm}
\title{This function returns the normalizing factor, based upon the input settings}
\usage{
getNormalizationTerm(dates, flags, defaultTime = 1)
}
\arguments{
\item{dates}{The dates of all visits recorded in the record}

\item{flags}{The R dataframe that contains all feature/model flags
specified in settings.R. - specifies which sort of normalization to perform}

\item{defaultTime}{Value by which to normalize patients who only have a single visit, so cannot say what the follow-up time is (=0 --> undefined).  Set default time as 1; as if spreading single observation over an entire year or 1 month (depending on settings)}
}
\value{
The value by which to divide term counts
}
\description{
This function returns the normalizing factor, based upon the input settings
}
\details{
Depending upon the input settings, will return the normalizing term for the feature values.  Some normalization settings depend upon the specific feature type; these are addressed within the individual feature types.  This is a helper function for getPatientData
}
\examples{
\dontrun{

 timeDiff <- getNormalizationTerm(dates, flags)

}

}

