% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getPatientCohort_w_Anchors}
\alias{getPatientCohort_w_Anchors}
\title{This function builds a patient cohort (and controls) based on Anchors and lists}
\usage{
getPatientCohort_w_Anchors(connection, dbms, includeConceptlist,
  excludeConceptlist, schema, cohortSize, controlSize, flags)
}
\arguments{
\item{connection}{The connection to the database server.}

\item{dbms}{The target DBMS for SQL to be rendered in.}

\item{includeConceptlist}{The list of concept_id's used to build the
cohort.}

\item{excludeConceptlist}{The list of concept_id's used as exclusion
criteria for the cohort.}

\item{schema}{The database schema being used.}

\item{cohortSize}{The number of desired patients to appear in the cohort.}

\item{controlSize}{The number of desired patients to be in the control
group.}

\item{flags}{The flags variable containg the study configuration - we
use this one here to have flexibility of having two different sets of settings
for the same experimental run.}
}
\value{
A list of dataframes containing both cases and control patient_id's.
}
\description{
This function will build a patient cohort with its respective
controls using an inclusion concept_id list as well as an exclussion
concept_id list. The user specifies the number of both cases and controls for
his cohort.
}
\details{
This function takes the lists of include and exclude concept_ids and
finds all patients that satisfy this characteristics from the Observation and
Condition_occurrence tables in CDM V5.
}
\examples{
\dontrun{

casesANDcontrolspatient_ids_df<- getPatientCohort_w_Anchors(conn, dbms,
         as.character(keywordList_FF$V3), as.character(ignoreList_FF$V3),
         cdmSchema,nCases,nControls, flag)
if (nCases > nrow(casesANDcontrolspatient_ids_df[[1]])) {
     message("Not enough patients to get the number of cases specified")
     stop
} else {
   if (nCases > nrow(casesANDcontrolspatient_ids_df[[2]])) {
       message("Not enough patients to get the number of controls specified")
       stop
   }
}

}

}
