% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/functions.R
\name{getPatientData}
\alias{getPatientData}
\title{This function fetches all the patient data (generic)}
\usage{
getPatientData(connection, dbms, patient_ids, flags, schema)
}
\arguments{
\item{connection}{The connection to the database server.}

\item{dbms}{The target DBMS for SQL to be rendered in.}

\item{patient_ids}{The list of case patient id's to extract data from.}

\item{flags}{The R dataframe that contains all feature/model flags
specified in settings.R.}

\item{schema}{The database schema being used.}
}
\value{
An object containing the raw feature sets for the patient data.
}
\description{
This function fetches all the patient data (generic). Returns
raw patient data.
}
\details{
Based on the groups of feature sets determined in the flags
variable, this function will fetch patient data. The function returns all
patient information
}
\examples{
\dontrun{

 dataFcontrols <- getPatientData(conn, dbms, controls, flag , cdmSchema)

}
}

