% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{getPatientDataCases}
\alias{getPatientDataCases}
\title{This function fetches all the patient data (non-generic) designed to work
when building a model}
\usage{
getPatientDataCases(connection, dbms, patient_ids, keywords, ignores, flags,
  schema, removeDomains = "")
}
\arguments{
\item{connection}{The connection to the database server.}

\item{dbms}{The target DBMS for SQL to be rendered in.}

\item{patient_ids}{The list of case patient id's to extract data from.}

\item{keywords}{The list of concept_id's used to build the cohort.}

\item{ignores}{The list of concept_id's ignored when building the
cohort.}

\item{flags}{The R dataframe that contains all feature/model flags
specified in settings.R.}

\item{schema}{The database schema being used.}

\item{removeDomains=''}{List of domains to not include as features, if any are specified in settings file}
}
\value{
An object containing the raw feature sets for the patient data.
}
\description{
This function fetches all the patient data (non-generic)
designed to work when building a model. Returns raw patient data.
}
\details{
Based on the groups of feature sets determined in the flags
variable, this function will fetch patient data. The function determines the
first mention of the keywords and selects that date to start the data
extraction of the remaining patient information
}
\examples{
\dontrun{

 dataFcases <-getPatientDataCases(conn, dbms, cases, as.character(keywordList_FF$V3),
         flag , cdmSchema)

}
}

