% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getPatientDataStartEnd}
\alias{getPatientDataStartEnd}
\title{This function fetches all the patient data (generic) - from a given start date and with a given end date}
\usage{
getPatientDataStartEnd(connection, dbms, patient_ids, startDate, endDate,
  flags, schema, removeDomains = c(""))
}
\arguments{
\item{connection}{The connection to the database server.}

\item{dbms}{The target DBMS for SQL to be rendered in.}

\item{patient_ids}{The list of case patient id's to extract data from - NOT a data.frame.}

\item{startDate}{The start index date for all patients}

\item{endDate}{The end date to fetch data from patients}

\item{flags}{The R dataframe that contains all feature/model flags
specified in settings.R.}

\item{schema}{The database schema being used.}

\item{removeDomains=''}{List of domains to not include as features, if any are specified in settings file}
}
\value{
An object containing the raw feature sets for the patient data.
}
\description{
This function fetches all the patient data (generic). Returns
raw patient data.
}
\details{
Based on the groups of feature sets determined in the flags
variable, this function will fetch patient data within the specified
time range the function returns all patient information
}
\examples{
\dontrun{

 patientData <- getPatientDataStartEnd(conn, dbms, patient_ids, start_dates, end_dates, flag , cdmSchema)

}

}
