% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plotFeatWeightings}
\alias{plotFeatWeightings}
\title{This function plots the feature importance weightings}
\usage{
plotFeatWeightings(plotSaveFile, weightingsDF)
}
\arguments{
\item{plotSaveFile}{The name of the file to save}

\item{weightingsDF}{Data frame of the weightings with their labels}
}
\value{
(none)
}
\description{
This function plots the feature importance weightings
}
\details{
This function returned predicted classes for the input patient list.  Use case: evaluate trained model on a set of gold-standard patients.
}
\examples{
\dontrun{

 plotFeatWeightings(plotSaveFile, weightingsDF)

}

}
