% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{runGUI}
\alias{runGUI}
\title{This function runs the shiny app}
\usage{
runGUI()
}
\arguments{
\item{example}{The name of the embedded app found in the package.}
}
\value{
Nothing
}
\description{
This function runs the shiny app
}
\details{
This function is in charge of running the Aphrodite GUI
}
\examples{
\dontrun{

 runGUI("Aphrodite")

}

}
